Move config and logger to root of api dir. Add migration dir to config.
